services/loader-api-trades/requirements.txt
aiohttp
aiokafka
loguru
uvloop
python-dotenv
